requirements.txt
python-telegram-bot==21.5
python-dotenv
SQLAlchemy==2.0.34
